combine everything

+++ libs/genetic-algorithm/src/chromosome.rs
@@ -3,7 +3,7 @@ use std::ops::Index;
 
 #[derive(Clone, Debug)]
 pub struct Chromosome {
-    genes: Vec<f32>,
+    genes: Vec<f64>,
 }
 
 #[allow(clippy::len_without_is_empty)] // chromosomes are not supposed to be empty, so it doesn't make much sense to have such function
@@ -12,25 +12,25 @@ impl Chromosome {
         self.genes.len()
     }
 
-    pub fn iter(&self) -> impl Iterator<Item = &f32> {
+    pub fn iter(&self) -> impl Iterator<Item = &f64> {
         self.genes.iter()
     }
 
-    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut f32> {
+    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut f64> {
         self.genes.iter_mut()
     }
 }
 
 impl Index<usize> for Chromosome {
-    type Output = f32;
+    type Output = f64;
 
     fn index(&self, index: usize) -> &Self::Output {
         &self.genes[index]
     }
 }
 
-impl FromIterator<f32> for Chromosome {
-    fn from_iter<T: IntoIterator<Item = f32>>(iter: T) -> Self {
+impl FromIterator<f64> for Chromosome {
+    fn from_iter<T: IntoIterator<Item = f64>>(iter: T) -> Self {
         Self {
             genes: iter.into_iter().collect(),
         }
@@ -38,8 +38,8 @@ impl FromIterator<f32> for Chromosome {
 }
 
 impl IntoIterator for Chromosome {
-    type Item = f32;
-    type IntoIter = impl Iterator<Item = f32>;
+    type Item = f64;
+    type IntoIter = impl Iterator<Item = f64>;
 
     fn into_iter(self) -> Self::IntoIter {
         self.genes.into_iter()
